+++ src/scyjava/_introspect.py
@@ -3,8 +3,7 @@ Introspection functions for reporting Java
 class methods, fields, and source code URL.
 """
 
-from functools import partial
-from typing import Any, Dict, List, Optional
+from typing import Any, Dict, List
 
 from scyjava._jvm import jimport
 from scyjava._types import isjava, jinstance, jclass
@@ -64,7 +63,7 @@ def jreflect(data, aspect: str = "all") -> List[Dict[str, Any]]:
         returns = (
             member.getReturnType().getName()
             if hasattr(member, "getReturnType")
-            else (member.getType().getName() if hasattr(member, "getType") else None)
+            else (member.getType().getName() if hasattr(member, "getType") else name)
         )
         table.append(
             {
@@ -77,58 +76,6 @@ def jreflect(data, aspect: str = "all") -> List[Dict[str, Any]]:
         )
 
     return table
-
-
-def _map_syntax(base_type):
-    """
-    Map a Java BaseType annotation (see link below) in an Java array
-    to a specific type with an Python interpretable syntax.
-    https://docs.oracle.com/javase/specs/jvms/se7/html/jvms-4.html#jvms-4.3
-    """
-    basetype_mapping = {
-        "[B": "byte[]",
-        "[C": "char[]",
-        "[D": "double[]",
-        "[F": "float[]",
-        "[I": "int[]",
-        "[J": "long[]",
-        "[L": "[]",  # array
-        "[S": "short[]",
-        "[Z": "boolean[]",
-    }
-
-    if base_type in basetype_mapping:
-        return basetype_mapping[base_type]
-    # Handle the case of a returned array of an object
-    elif base_type.__str__().startswith("[L"):
-        return base_type.__str__()[2:-1] + "[]"
-    else:
-        return base_type
-
-
-def _make_pretty_string(entry, offset):
-    """
-    Print the entry with a specific formatting and aligned style.
-    :param entry: Dictionary of class names, modifiers, arguments, and return values.
-    :param offset: Offset between the return value and the method.
-    """
-
-    # A star implies that the method is a static method
-    return_val = f"{entry['returns'].__str__():<{offset}}"
-    # Handle whether to print static/instance modifiers
-    obj_name = f"{entry['name']}"
-    modifier = f"{'*':>4}" if "static" in entry["mods"] else f"{'':>4}"
-
-    # Handle fields
-    if entry["arguments"] is None:
-        return f"{return_val} {modifier} = {obj_name}\n"
-
-    # Handle methods with no arguments
-    if len(entry["arguments"]) == 0:
-        return f"{return_val} {modifier} = {obj_name}()\n"
-    else:
-        arg_string = ", ".join([r.__str__() for r in entry["arguments"]])
-        return f"{return_val} {modifier} = {obj_name}({arg_string})\n"
 
 
 def jsource(data):
@@ -162,66 +109,3 @@ def jsource(data):
         return f"Not a Java class {str(type(data))}"
     except Exception as err:
         return f"Unexpected {err=}, {type(err)=}"
-
-
-def _print_data(data, aspect, static: Optional[bool] = None, source: bool = True):
-    """
-    Write data to a printed string of class methods with inputs, static modifier,
-    arguments, and return values.
-
-    :param data: The object or class to inspect or fully qualified class name.
-    :param aspect: Whether to print class "fields" or "methods".
-    :param static:
-        Boolean filter on Static or Instance methods.
-        Optional, default is None (prints all).
-    :param source: Whether to print any available source code. Default True.
-    """
-    table = jreflect(data, aspect)
-    if len(table) == 0:
-        print(f"No {aspect} found")
-        return
-
-    # Print source code
-    offset = max(list(map(lambda entry: len(entry["returns"]), table)))
-    all_methods = ""
-    if source:
-        urlstring = jsource(data)
-        print(f"Source code URL: {urlstring}")
-
-    # Print methods
-    for entry in table:
-        entry["returns"] = _map_syntax(entry["returns"])
-        if entry["arguments"]:
-            entry["arguments"] = [_map_syntax(e) for e in entry["arguments"]]
-        if static is None:
-            entry_string = _make_pretty_string(entry, offset)
-            all_methods += entry_string
-
-        elif static and entry["static"]:
-            entry_string = _make_pretty_string(entry, offset)
-            all_methods += entry_string
-        elif not static and not entry["static"]:
-            entry_string = _make_pretty_string(entry, offset)
-            all_methods += entry_string
-        else:
-            continue
-
-    # 4 added to align the asterisk with output.
-    print(f"{'':<{offset + 4}}* indicates static modifier")
-    print(all_methods)
-
-
-# The functions with short names for quick usage.
-methods = partial(_print_data, aspect="methods")
-fields = partial(_print_data, aspect="fields")
-attrs = partial(_print_data, aspect="fields")
-
-
-def src(data):
-    """
-    Print the source code URL for a Java class, object, or class name.
-
-    :param data: The Java class, object, or fully qualified class name as string
-    """
-    source_url = jsource(data)
-    print(f"Source code URL: {source_url}")
